15_autogen_microsoft/04_agents_messages/agents_messages/pyproject.toml
[project]
name = "agents-messages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Zia Khan", email = "ziaukhan@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.4.7",
    "autogen-ext[openai]>=0.4.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20241016",
]

[project.scripts]
textmessage = "agents_messages.message:run_main"
multimodal = "agents_messages.multimodal:run_main"
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
